Prevent IrreversibleMigration exception
Raising an exception during the down phase of a DB migration which
operationally precedes a drop_table anyway simply blocks removal
of the plugin. Log the message and present to the user, and since
in most cases the plugin is being removed when this is called, it
will not affect any relevant data or operations.

This is not an ideal solution, but a monkey-patch around rake's
handling of exceptions raised during chained task execution.